### 2023-02-21 #5180 @editorialbot
[PRE REVIEW]: HRnV-Calc: A Software for Heart Rate n-Variability and Heart Rate Variability Analysis
**Submitting author:** @charlienew (<a href="http://orcid.org/0000-0001-6677-7910">Chenglin Niu</a>)
**Repository:** https://github.com/nliulab/HRnV-Calc
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
