### 2023-04-17 #5180 @Kevin-Mattheus-Moerman
@bmcfee thanks for handling this submission. Do you think you can work on recruiting some more reviewers? There are a number of accepted papers in JOSS on heart rate analysis, perhaps authors/reviewers for those can help too. Thanks again.
