### 2023-04-28 nliulab/HRnV-Calc#3 @smnnlt
Add contributing guidelines
Please add guidelines for users on how to contribute to your software and how to seek help (see https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines). This may also include adding a code of conduct.

Issue during review for the JOSS (https://github.com/openjournals/joss-reviews/issues/5391)
