### 2023-04-28 nliulab/HRnV-Calc#4 @smnnlt
Installation instructions for non-git users
The aim of your software with the GUI-interface is to reach a wide audience including non-programmers. Currently the README only provides installation instruction using git. To enable users not using git to use your software I recommend to include a comment on how to install the software without using git. This may be as simple as pointing out to the "Download Zip" option on Github.

Issue during review for the JOSS (https://github.com/openjournals/joss-reviews/issues/5391).
