### 2023-04-28 nliulab/HRnV-Calc#5 @smnnlt
Treat file extensions as default postfix during patient id extraction
I don't think there is any case in which you want the file extension of the raw data file to be part of the new file names/extracted patient IDs. Thus I suggest to automatically remove the file extension for the patient id extraction, or in other words to add the file extension as the default postfix.

Issue during review for the JOSS (https://github.com/openjournals/joss-reviews/issues/5391).
