### 2023-04-28 nliulab/HRnV-Calc#6 @smnnlt
Consider automated testing
Thanks to the detailed README, I could reproduce the results validation for the provided raw data. However, this does not included the validation of non-default processing options and the treatment of other input files. I highly encourage the authors to use unit testing (https://de.mathworks.com/help/matlab/matlab_prog/ways-to-write-unit-tests.html) to ensure the validity of their work, for example by testing against the PhysioNet toolbox or during further development against previous versions of this software.

Issue during review for the JOSS (https://github.com/openjournals/joss-reviews/issues/5391).
