### 2023-04-28 #5391 @smnnlt
Hi @charlienew , here are some additional comments to my review:

First of all thank you very much for providing this interesting piece of software. I think the GUI interface for HR(n)V analysis may be very helpful in practical settings and I look forward to future insights into the newly developed measurement parameters. In addition to the checklist above and the issues I opened in the repo I have some remarks regarding the software paper:

**Authorship**: From what I can see, three of the ten authors listed on the paper have contributed code to the software. While it is ultimately your choice who to add as an author, please double-check that all other people have made non-code contributions specifically to this software (e.g., software testing, conceptualization).

**Summary**: I appreciate your efforts in providing a well-referenced introduction to HR(n)V analysis in the introduction of your paper. Even for me as a sport scientist (though not working in the field of HRV research) it was indeed hard to follow the technical parts of this section. Please note, that the summary section of the paper should avoid jargon and be adressed to a non-specialist audience (https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain).

**Basic Usage**: I believe this section could be much shorter. The quite detailed API walkthrough is mostly already part of your well-written README and should not be part of the JOSS paper. However I think that showing some high-level functionality of the software may also be useful in the paper, so I suggest to keep Figure 1 with the general workflow and to shortly mention important other features without going much into detail. This could also help you to archieve the target word count for the paper (<1000). Additionally, there are some references in this section related to specific processing options that should rather be part of the software documentation (i.e., the README in your case).
