### 2023-05-02 #5391 @charlienew
Thank you @smnnlt and @itismeghasyam for reviewing our software and the constructive comments.

To respond to the comments from @smnnlt, first of all, thank you for pointing out the concern for authorship. We know that the list might be longer than most papers on JOSS. However, we can confirm that all other authors listed have contributed to our software either in conceptualization or in testing. Although the code was only written by three people, the design and the workflow of HRnV-Calc are the results of wide collaboration between clinicians and HRV researchers.

We understand that the introduction of our paper might be somewhat lengthy and exhaustive. However, we do think a bit of technical details is necessary since the HRnV method is quite new, and not many people are familiar with its background and the motives for its introduction. We believe most of the potential users of HRnV-Calc to be in the field of HRV and related research, in particular, clinicians dealing with cardiac-related conditions. Therefore, some in-depth descriptions and discussions of HRnV would help them to decide whether to try our software. However, we will work on the writing of this section to make it more appealing to a broader audience.

The comments on the basic usage section are absolutely right. We will remove all the figures except Figure 1 and shorten the descriptions for each GUI.

We will also address the issues raised in our HRnV-Calc repo and respond directly in the corresponding issue threads. Once all necessary changes are done, we will update our code and paper in the repo. Thanks again for your feedback and suggestions!
