### 2023-05-04 #5391 @charlienew
Hi @itismeghasyam, thanks for your suggestions. Yes, we will include a section in our README about contributing guidelines and a code of conduct, which should be completed by the end of this week along with some other issues raised by @smnnlt.
