### 2023-05-10 4376898
Feature and README Update
Feature update:

Patient ID will no longer include file extensions by default. (See issue #5)

README updates:

1. Added instructions of download for non-git users. (See issue #4)

2. Added contribution guidelines and support instructions. (See issue #3)
