+++ paper/paper.bib
@@ -24,41 +24,6 @@ Reproducibility of Results},
    type = {Journal Article}
 }
 
-@article{RN5,
-   author = {Behar, J. and Johnson, A. and Clifford, G. D. and Oster, J.},
-   title = {A comparison of single channel fetal ECG extraction methods},
-   journal = {Ann Biomed Eng},
-   volume = {42},
-   number = {6},
-   pages = {1340-53},
-   note = {Behar, Joachim
-Johnson, Alistair
-Clifford, Gari D
-Oster, Julien
-eng
-098461/Z/12/Z/Wellcome Trust/United Kingdom
-Comparative Study
-Research Support, Non-U.S. Gov't
-Ann Biomed Eng. 2014 Jun;42(6):1340-53. doi: 10.1007/s10439-014-0993-9. Epub 2014 Mar 7.},
-   abstract = {The abdominal electrocardiogram (ECG) provides a non-invasive method for monitoring the fetal cardiac activity in pregnant women. However, the temporal and frequency overlap between the fetal ECG (FECG), the maternal ECG (MECG) and noise results in a challenging source separation problem. This work seeks to compare temporal extraction methods for extracting the fetal signal and estimating fetal heart rate. A novel method for MECG cancelation using an echo state neural network (ESN) based filtering approach was compared with the least mean square (LMS), the recursive least square (RLS) adaptive filter and template subtraction (TS) techniques. Analysis was performed using real signals from two databases composing a total of 4 h 22 min of data from nine pregnant women with 37,452 reference fetal beats. The effects of preprocessing the signals was empirically evaluated. The results demonstrate that the ESN based algorithm performs best on the test data with an F1 measure of 90.2% as compared to the LMS (87.9%), RLS (88.2%) and the TS (89.3%) techniques. Results suggest that a higher baseline wander high pass cut-off frequency than traditionally used for FECG analysis significantly increases performance for all evaluated methods. Open source code for the benchmark methods are made available to allow comparison and reproducibility on the public domain data.},
-   keywords = {*Algorithms
-*Databases, Factual
-Electrocardiography/*methods
-Female
-Fetus/*physiology
-Heart Rate/*physiology
-Humans
-*Models, Theoretical
-Pregnancy
-*Signal Processing, Computer-Assisted},
-   ISSN = {1573-9686 (Electronic)
-0090-6964 (Linking)},
-   DOI = {10.1007/s10439-014-0993-9},
-   url = {https://www.ncbi.nlm.nih.gov/pubmed/24604619},
-   year = {2014},
-   type = {Journal Article}
-}
-
 @article{RN88,
    author = {Bohanon, F. J. and Mrazek, A. A. and Shabana, M. T. and Mims, S. and Radhakrishnan, G. L. and Kramer, G. C. and Radhakrishnan, R. S.},
    title = {Heart rate variability analysis is more sensitive at identifying neonatal sepsis than conventional vital signs},
@@ -412,8 +377,8 @@ Middle Aged},
    type = {Journal Article}
 }
 
-@article{RN55,
-   author = {Liu, C. and Li, K. and Zhao, L. and Liu, F. and Zheng, D. and Liu, C. and Liu, S.},
+@article{Liu:2013,
+   author = {{Liu}, C. and Li, K. and Zhao, L. and Liu, F. and Zheng, D. and Liu, C. and Liu, S.},
    title = {Analysis of heart rate variability using fuzzy measure entropy},
[…]
