+++ paper/paper.md
@@ -29,6 +29,7 @@ authors:
   - name: Gari D. Clifford
     affiliation: "6,7"
   - name: Nan Liu
+    orcid: 0000-0003-3610-4883
     corresponding: true 
     affiliation: "1,8,9"
 
@@ -47,7 +48,7 @@ affiliations:
    index: 6
  - name: Department of Biomedical Engineering, Georgia Institute of Technology, Atlanta, GA, United States of America
    index: 7
- - name: SingHealth AI Health Program, Singapore Health Services, Singapore, Singapore 
+ - name: SingHealth AI Office, Singapore Health Services, Singapore, Singapore 
    index: 8
  - name: Institute of Data Science, National University of Singapore, Singapore, Singapore
    index: 9
@@ -60,33 +61,30 @@ bibliography: paper.bib
 
 Variation of the time interval between a consistent point in time of each heartbeat (generally related to ventricular electrical activation), known as heart rate variability (HRV) [@RN28], has been proven by numerous studies to be a useful indicator of physiological status [@RN69; @RN28]. Thanks to its non-invasive nature and strong connection to the autonomic nervous system (ANS) [@RN69; @RN28], HRV has been adopted to study a wide range of diseases and clinical conditions, which include myocardial infarction [@RN83], sudden cardiac death [@RN84], diabetes [@RN86], renal failure [@RN87], sepsis [@RN88], seizure [@RN97], and cancer [@RN90]. In addition, the emergence of wearable devices with heart monitoring capabilities has also allowed researchers to study the above-mentioned medical conditions in real-world settings [@RN91], as well as in non-clinical applications such as sports [@RN51], stress [@RN93], and sleep monitoring [@RN94]. 
 
-A variety of HRV metrics can be derived from the heartbeat time sequence known as the inter-beat interval (IBI), or the R-to-R peak interval (RR interval or RRI). Such sequences are often extracted from biomedical signals such as electrocardiograms (ECG) and photoplethysmograms (PPG). It is believed that a decrease of complexity in HRV is associated with an increase in both morbidity and mortality [@RN69; @RN28]. To qualify and quantify the complexity, various conventional HRV metrics in linear and non-linear domains [@RN23] have been established to reflect the dynamic of HRV [@RN28; @RN8]. However, the exact mechanism regulating HRV is not perfectly clear in every detail[@RN69; @RN28]. Some of recent developments in HRV have been mainly focused on non-linear metrics such as variants of approximate entropy (ApEn) and sample entropy (SampEn) [@RN95; @RN96; @RN82; @RN55]. 
+A variety of HRV metrics can be derived from the heartbeat time sequence known as the inter-beat interval (IBI), or the R-to-R peak interval (RR interval or RRI). Such sequences are often extracted from biomedical signals such as electrocardiograms (ECG) and photoplethysmograms (PPG). It is believed that a decrease of complexity in HRV is associated with an increase in both morbidity and mortality [@RN69; @RN28]. To qualify and quantify the complexity, various conventional HRV metrics in linear and non-linear domains [@RN23] have been established to reflect the dynamic of HRV [@RN28; @RN8]. However, the exact mechanism regulating HRV is not perfectly clear in every detai [@RN69; @RN28]. Some of recent developments in HRV have been mainly focused on nonlinear metrics such as variants of approximate entropy (ApEn) and sample entropy (SampEn) [@RN95; @RN96; @RN82; @Liu:2013]. 
 
[…]
