+++ paper/paper.bib
@@ -378,7 +378,7 @@ Middle Aged},
 }
 
 @article{Liu:2013,
-   author = {{Liu}, C. and Li, K. and Zhao, L. and Liu, F. and Zheng, D. and Liu, C. and Liu, S.},
+   author = {{Liu}, Chengyu. and Li, K. and Zhao, L. and Liu, F. and Zheng, D. and Liu, C. and Liu, S.},
    title = {Analysis of heart rate variability using fuzzy measure entropy},
    journal = {Comput Biol Med},
    volume = {43},
@@ -421,7 +421,7 @@ Statistics, Nonparametric},
 }
 
 @article{Liu:2020,
-   author = {{Liu}, N. and Guo, D. and Koh, Z. X. and Ho, A. F. W. and Xie, F. and Tagami, T. and Sakamoto, J. T. and Pek, P. P. and Chakraborty, B. and Lim, S. H. and Tan, J. W. C. and Ong, M. E. H.},
+   author = {{Liu}, Nan and Guo, D. and Koh, Z. X. and Ho, A. F. W. and Xie, F. and Tagami, T. and Sakamoto, J. T. and Pek, P. P. and Chakraborty, B. and Lim, S. H. and Tan, J. W. C. and Ong, M. E. H.},
    title = {Heart rate n-variability (HRnV) and its application to risk stratification of chest pain patients in the emergency department},
    journal = {BMC Cardiovasc Disord},
    volume = {20},
@@ -477,7 +477,7 @@ Time Factors
 }
 
 @article{Liu:2021,
-   author = {{Liu}, N. and Chee, Marcel Lucas and Foo, Mabel Zhi Qi and Pong, Jeremy Zhenwen and Guo, Dagang and Koh, Zhi Xiong and Ho, Andrew Fu Wah and Niu, Chenglin and Chong, Shu-Ling and Ong, Marcus Eng Hock},
+   author = {{Liu}, Nan and Chee, Marcel Lucas and Foo, Mabel Zhi Qi and Pong, Jeremy Zhenwen and Guo, Dagang and Koh, Zhi Xiong and Ho, Andrew Fu Wah and Niu, Chenglin and Chong, Shu-Ling and Ong, Marcus Eng Hock},
    title = {Heart rate n-variability (HRnV) measures for prediction of mortality in sepsis patients presenting at the emergency department},
    journal = {PLOS ONE},
    volume = {16},
