+++ paper/paper.bib
@@ -378,7 +378,7 @@ Middle Aged},
 }
 
 @article{Liu:2013,
-   author = {{Liu}, Chengyu. and Li, K. and Zhao, L. and Liu, F. and Zheng, D. and Liu, C. and Liu, S.},
+   author = {{Liu}, Chengyu and Li, K. and Zhao, L. and Liu, F. and Zheng, D. and Liu, C. and Liu, S.},
    title = {Analysis of heart rate variability using fuzzy measure entropy},
    journal = {Comput Biol Med},
    volume = {43},
