### 2023-05-12 8928bce
Spelling corrections in README
