+++ README.md
@@ -281,8 +281,7 @@ As shown in figure above, the HR<sub>n</sub>V<sub>m</sub> Results Display viewer
 
 ### Example: Batch Processing
 
-HRnV-Calc supports analyzing multiple RRI signals simultaneously as batch processing.  [Three demo RRI inputs]((./Demo_Data/Demo_RRI/)) are provided in this repository. To start, check the `Batch Files` option in [Data Loader](#data-loader) as shown in the figure below. The `Prefix` and `Postfix` sections can be specified for [patient ID extractions](#patient-id-extraction). Do note that the 
-[HR<sub>n</sub>V<sub>m</sub> Results Display](#hrnvm-results-display) will not be displayed for batch processing. 
+HRnV-Calc supports analyzing multiple RRI signals simultaneously as batch processing.  [Three demo RRI inputs]((./Demo_Data/Demo_RRI/)) are provided in this repository. To start, check the `Batch Files` option in [Data Loader](#data-loader) as shown in the figure below. The `Prefix` and `Postfix` sections can be specified for [patient ID extractions](#patient-id-extraction). Do note that the [HR<sub>n</sub>V<sub>m</sub> Results Display](#hrnvm-results-display) will not be displayed for batch processing.
 <p align="center">
   <img src="./figs/Batch_processing_data_loader.png" />
 </p>
@@ -290,7 +289,7 @@ HRnV-Calc supports analyzing multiple RRI signals simultaneously as batch proces
   Data Loader Settings for Batch Processing
 </p>
 
-The configuration of a single HR<sub>n</sub>V<sub>m</sub> of all the input RRI files is the same as mentioned in [HR<sub>n</sub>V<sub>m</sub> Setting Viewer](#hrnvm-setting-viewer). Here, we show the settings of multiple HR<sub>n</sub>V<sub>m</sub> analyses by choosing the option `All` in the figure below. 
+The configuration of a single HR<sub>n</sub>V<sub>m</sub> of all the input RRI files is the same as mentioned in [HR<sub>n</sub>V<sub>m</sub> Setting Viewer](#hrnvm-setting-viewer). Here, we show the settings of multiple HR<sub>n</sub>V<sub>m</sub> analyses by choosing the option `All` in the figure below.
 
 <p align="center">
   <img src="./figs/Batch_processing_hrnvm_settings.png" />
@@ -300,7 +299,7 @@ The configuration of a single HR<sub>n</sub>V<sub>m</sub> of all the input RRI f
 </p>
 
 
-In this example, we choose `n` number to be 3, meaning all HR<sub>n</sub>V<sub>m</sub> analyses with `n` and `m` number smaller or equal to 3 will be performed (6 different analyses in total). All other settings, such  as `Ectopic Beats` and `Frequency Domain` settings are set to default. The reults of these anlyses can be found in the [demo results folder](Demo/Demo_Results/RRI_Results/).
+In this example, we choose `n` number to be 3, meaning all HR<sub>n</sub>V<sub>m</sub> analyses with `n` and `m` number smaller or equal to 3 will be performed (6 different analyses in total). All other settings, such  as `Ectopic Beats` and `Frequency Domain` settings are set to default. The results of these analyses can be found in the [demo results folder](Demo/Demo_Results/RRI_Results/).
 
 ## Results Verification 
 
@@ -308,19 +307,19 @@ To help users determine the integrity of their installation, we provide some HR<
 
 The [ECG results](/Demo/Demo_Results/ECG_Results/) were obtained using the [single demo ECG file](Demo/Demo_Data/Demo_ECG/Demo_NSR16786.txt) (sampling rate: 128Hz). There are four files under the this folder:
 
[…]
