### 2023-05-12 nliulab/HRnV-Calc#6 @charlienew
Thanks @smnnlt  for your valuable suggestion. It is important to have exhaustive unit testing for our software. However, our software is unique as it consists of manual annotation component in addition to automatic parameter calculation. Therefore, implementing unit tests could be complicated and may take a considerable amount of effort in planning and execution. We fully agree with your suggestion and plan to gradually roll out these tests in future versions of HRnV-Calc.
