### 2023-05-12 nliulab/HRnV-Calc#5 @charlienew
Thanks @smnnlt for your suggestion. We have updated the software to fix this issue in our latest release.
