### 2023-05-12 nliulab/HRnV-Calc#4 @charlienew
Thanks @smnnlt for your suggestion. We added additional instructions in README about how to download our software directly from GitHub. Moreover, we also plan to release a compiled version of HRnV-Calc for people without Matlab installation in the future.
