### 2023-05-12 nliulab/HRnV-Calc#3 @charlienew
Thanks @smnnlt for your suggestion. We have added necessary guidelines in README about how others may be able to contribute to our software and how they could seek help on GitHub.
