### 2023-05-12 #5391 @charlienew
Hi @smnnlt and @itismeghasyam,

Thanks again for reviewing our software and paper. We have updated our code and GitHub page to address the valuable feedback and suggestions. For the issues raised on our GitHub page, we have addressed them individually in the issue thread and made necessary changes to our code to fix the issues.

We have also revised our paper to shorten the ‘Basic Usage’ section, which now only includes Figure 1 and some brief descriptions of each GUI.

Thank you.
