### 2023-05-15 #5391 @bmcfee
@smnnlt yes, I think the level of technical detail here is in keeping with the expectations outlined in the submission guidelines: https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain
