### 2023-05-19 #5391 @bmcfee
Thanks @itismeghasyam and @smnnlt !

@charlienew I've read through the paper, and there are a few things that should be corrected before the final version:

- Line 63: "these tool" → "these tools"
- Bibliography entry here https://github.com/nliulab/HRnV-Calc/blob/8928bceca1b25c31739212d78e8fbf9a8b87f469/paper/paper.bib#LL700C72-L700C72 is coming out badly formatted due to incorrect formatting of diacritics.  Can you fix this, eg, `´i` → `í` and `~n` → `ñ`?  (Or use latex/bibtex format if you prefer.)
- Bibliography entry here https://github.com/nliulab/HRnV-Calc/blob/8928bceca1b25c31739212d78e8fbf9a8b87f469/paper/paper.bib#L117 is breaking the last name `van Leeuwen` incorrectly; should be encoded as `{van Leeuwen}, P.`
- Bibliography entry here https://github.com/nliulab/HRnV-Calc/blob/8928bceca1b25c31739212d78e8fbf9a8b87f469/paper/paper.bib#L2 is also rendering strangely.  Suggestion is to add surrounding curly braces `{}` to the author entry, and remove the redundant copy of this info from the title field.  `url` field should be replaced by the DOI
- Please remove the `type` field from bibliography entries
