+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @article{RN8,
-   author = {Task Force of the European Society of Cardiology and the North American Society of Pacing and Electrophysiology.},
-   title = {Heart rate variability. Standards of measurement, physiological interpretation, and clinical use. Task Force of the European Society of Cardiology and the North American Society of Pacing and Electrophysiology},
+   author = {{Task Force of the European Society of Cardiology and the North American Society of Pacing and Electrophysiology}},
+   title = {Heart rate variability. Standards of measurement, physiological interpretation, and clinical use.},
    journal = {Eur Heart J},
    volume = {17},
    number = {3},
@@ -19,9 +19,8 @@ Myocardial Infarction/diagnosis
 Reproducibility of Results},
    ISSN = {0195-668X (Print)
 0195-668X (Linking)},
-   url = {https://www.ncbi.nlm.nih.gov/pubmed/8737210},
-   year = {1996},
-   type = {Journal Article}
+   url = {https://doi.org/10.1161/01.CIR.93.5.1043},
+   year = {1996}
 }
 
 @article{RN88,
@@ -35,8 +34,7 @@ Reproducibility of Results},
    ISSN = {0002-9610
 1879-1883},
    DOI = {10.1016/j.amjsurg.2015.06.002},
-   year = {2015},
-   type = {Journal Article}
+   year = {2015}
 }
 
 @article{RN83,
@@ -72,8 +70,7 @@ Time Factors},
    ISSN = {1128-3602 (Print)
 1128-3602 (Linking)},
    url = {https://www.ncbi.nlm.nih.gov/pubmed/19694345},
-   year = {2009},
-   type = {Journal Article}
+   year = {2009}
 }
 
 @article{RN70,
@@ -109,12 +106,11 @@ Non-programmatic},
 0031-9007 (Linking)},
    DOI = {10.1103/PhysRevLett.89.068102},
    url = {https://www.ncbi.nlm.nih.gov/pubmed/12190613},
-   year = {2002},
-   type = {Journal Article}
+   year = {2002}
 }
 
 @article{RN62,
-   author = {Cysarz, D. and Bettermann, H. and van Leeuwen, P.},
+   author = {Cysarz, D. and Bettermann, H. and {van Leeuwen}, P.},
    title = {Entropies of short binary sequences in heart period dynamics},
    journal = {Am J Physiol Heart Circ Physiol},
    volume = {278},
@@ -140,8 +136,7 @@ Reproducibility of Results},
 0363-6135 (Linking)},
    DOI = {10.1152/ajpheart.2000.278.6.H2163},
    url = {https://www.ncbi.nlm.nih.gov/pubmed/10843917},
-   year = {2000},
-   type = {Journal Article}
+   year = {2000}
 }
 
 @article{RN100,
@@ -185,8 +180,7 @@ support vector machine (SVM)},
 1424-8220 (Linking)},
    DOI = {10.3390/s20030765},
    url = {https://www.ncbi.nlm.nih.gov/pubmed/32019220},
-   year = {2020},
-   type = {Journal Article}
+   year = {2020}
 }
 
 @article{RN99,
@@ -199,8 +193,7 @@ support vector machine (SVM)},
[…]
