+++ paper/paper.md
@@ -67,7 +67,7 @@ Despite progress in HRV metrics research, the representation of RRI upon which H
 
 # Statement of need
 
-There is an abundance of HRV software tools for commercial and non-commercial use, including Kubios HRV [@RN3], ECGlab [@RN66], ARTiiFACT [@RN65], RHRV [@RN64], and RR-APET [@RN72]. However, none of these tools is suitable for incorporating HRnV analysis. Moreover, these tool provide inconsistent results, making comparisons between research impossible [@RN4]. Since HRnV shares some common processing methods with conventional HRV analysis, it is natural to develop the HRnV package based on existing benchmarked software. We therefore developed an open-source HRnV software, HRnV-Calc, based on the PhysioNet Cardiovascular Signal Toolbox (PCST) [@RN4]. Compared to other HRV freeware, the PCST is standardized and well-documented. More importantly, the PCST is an open-source HRV software suite which has gone through rigorous testing and benchmarking in both technical and clinical settings. Based on the fully functional HRV command-line code provided by the PCST, HRnV-Calc has integrated graphical user interfaces (GUIs) that enable manual inspection and correction of RRI extraction from ECG signals, flexible configuration, and batch-processing, in a step-by-step manner. Its inherent functions support the analysis of both HRnV and conventional HRV metrics with enhanced usability. Therefore, HRnV-Calc not only facilitates new methodological developments, but also provide clinicians and researchers with transparent and easily accessible HRnV and HRV analyses.
+There is an abundance of HRV software tools for commercial and non-commercial use, including Kubios HRV [@RN3], ECGlab [@RN66], ARTiiFACT [@RN65], RHRV [@RN64], and RR-APET [@RN72]. However, none of them is suitable for incorporating HRnV analysis. Moreover, these tools provide inconsistent results, making comparisons between research impossible [@RN4]. Since HRnV shares some common processing methods with conventional HRV analysis, it is natural to develop the HRnV package based on existing benchmarked software. We therefore developed an open-source HRnV software, HRnV-Calc, based on the PhysioNet Cardiovascular Signal Toolbox (PCST) [@RN4]. Compared to other HRV freeware, the PCST is standardized and well-documented. More importantly, the PCST is an open-source HRV software suite which has gone through rigorous testing and benchmarking in both technical and clinical settings. Based on the fully functional HRV command-line code provided by the PCST, HRnV-Calc has integrated graphical user interfaces (GUIs) that enable manual inspection and correction of RRI extraction from ECG signals, flexible configuration, and batch-processing, in a step-by-step manner. Its inherent functions support the analysis of both HRnV and conventional HRV metrics with enhanced usability. Therefore, HRnV-Calc not only facilitates new methodological developments, but also provide clinicians and researchers with transparent and easily accessible HRnV and HRV analyses.
 
 # Basic Usage
 This section provides a non-exhaustive walkthrough of the features and functionalities offered by HRnV-Calc. 
