### 2023-05-22 #5391 @charlienew
Thanks again @itismeghasyam and @smnnlt for the review!

Thanks @bmcfee for your corrections. I have made the following changes to our paper:

- Line 63: I have replaced the first "these tools" with "them", and the second "these tool" was changed to "these tools" as you corrected.
- The name of Mr. Rodríguez-Liñares has been fixed. Thanks for your correction.
- The name of Mr. van Leeuwen has been fixed. Thanks for your correction.
- The name of the EU task force has been fixed. Thanks for the correction.
- All the `type` fields in the bibliography have been removed.
