+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @article{RN8,
-   author = {{Task Force of the European Society of Cardiology and the North American Society of Pacing and Electrophysiology}},
-   title = {Heart rate variability. Standards of measurement, physiological interpretation, and clinical use.},
+   author = {{Task Force of the European Society of Cardiology} and {The North American Society of Pacing and Electrophysiology}},
+   title = {Heart rate variability. Standards of measurement, physiological interpretation, and clinical use},
    journal = {Eur Heart J},
    volume = {17},
    number = {3},
