+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{RN8,
-   author = {{Task Force of the European Society of Cardiology} and {The North American Society of Pacing and Electrophysiology}},
+   author = {{Task Force of the European Society of Cardiology and the North American Society of Pacing and Electrophysiology}},
    title = {Heart rate variability. Standards of measurement, physiological interpretation, and clinical use},
    journal = {Eur Heart J},
    volume = {17},
