+++ paper/paper.bib
@@ -674,7 +674,7 @@ Risk Factors},
 }
 
 @article{RN64,
-   author = {Rodríguez-Liñares L, Vila X, Mendez A, Lado M, Olivieri D.},
+   author = {L. Rodríguez-Liñares, X.A. Vila, A. J. Méndez, M. J. Lado, and D. Olivieri},
    title = {RHRV: An R-based software package for heart rate variability analysis of ECG recordings},
    journal = {In 3rd Iberian Conference in Systems and Information Technologies (CISTI 2008)},
    year = {2008}
