+++ README.md
@@ -1,4 +1,5 @@
-# HRnV-Calc: A Software for Heart Rate n-Variability and Heart Rate Variability Analysis [![License: GPL v3](https://img.shields.io/badge/License-GPL%20v3-blue)](https://www.gnu.org/licenses/gpl-3.0.html)
+# HRnV-Calc: A Software for Heart Rate n-Variability and Heart Rate Variability Analysis 
+[![License: GPL v3](https://img.shields.io/badge/License-GPL%20v3-blue)](https://www.gnu.org/licenses/gpl-3.0.html)
  - **[HRnV-Calc Introduction](#hrnv-calc-introduction)**
       - [Description](#description)
       - [Citation](#citation)
