+++ paper/paper.bib
@@ -879,7 +879,9 @@ Reproducibility of Results
    title = {ECGlab: User friendly ECG/VCG analysis tool for research environments},
    booktitle = {Computing in Cardiology 2013},
    pages = {775-778},
-   ISBN = {2325-8853},
+   url = {http://www.cinc.org/archives/2013/pdf/0775.pdf},
+   biburl = {https://dblp.org/rec/conf/cinc/VicenteJGS13.bib},
+   bibsource = {dblp computer science bibliography, https://dblp.org},
    year = {2013}
 }
 
