### 2023-05-25 #5391 @charlienew
Hi @bmcfee, I just updated the BibTex entry for ECGLab. Hope this will fix the issue. Thanks.
