### 2023-05-25 #5391 @bmcfee
Thanks @charlienew - can you also remove the broken DOI For reference RN99? https://github.com/nliulab/HRnV-Calc/blob/fc6d7c2e58cb2b18c7c4ac969764fd8e37e6d9db/paper/paper.bib#L194
