+++ paper/paper.bib
@@ -191,7 +191,6 @@ support vector machine (SVM)},
    number = {3},
    pages = {289-295},
    abstract = {- Abstract: Introduction: Statistical markers are valuable when assessing physiological status over periods of time and in certain disease states. We assess if type 1 diabetes mellitus promote modification in autonomic nervous system using the main two types of algorithms to estimate a Fractal Dimension: Higuchi and Katz. Methods: 46 adults were divided into two equal groups. The autonomic evaluation consisted of recording heart rate variability (HRV) for 30 minutes in supine position in absence of any other stimuli. Fractal dimensions ought then able to determine which series of interbeat intervals are derived from diabetics’ or not. We then equated results to observe which assessment gave the greatest significance by One-way analysis of variance (ANOVA1), Kruskal-Wallis technique and Cohen’s d effect sizes. Results: Katz’s fractal dimension is the most robust algorithm when assisted by a cubic spline interpolation (6 Hz) to increase the number of samples in the dataset. This was categorical after two tests for normality; then, ANOVA1, Kruskal-Wallis and Cohen’s d effect sizes (p≈0.01 and Cohen’s d=0.814143 –medium effect size). Conclusion: Diabetes significantly reduced the chaotic response as measured by Katz’s fractal dimension. Katz’s fractal dimension is a viable statistical marker for subjects with type 1 diabetes mellitus.},
-   DOI = {-},
    url = {https://www.rjdnmd.org/index.php/RJDNMD/article/view/515},
    year = {2018}
 }
