### 2023-05-25 #5391 @bmcfee
My bad for not noticing the "-" DOI earlier :sweat_smile:

Looks like the bot has found a DOI For that reference though, so please add it back in and we should be all set.
