+++ paper/paper.bib
@@ -139,62 +139,6 @@ Reproducibility of Results},
    year = {2000}
 }
 
-@article{RN100,
-   author = {Czabanski, R. and Horoba, K. and Wrobel, J. and Matonia, A. and Martinek, R. and Kupka, T. and Jezewski, M. and Kahankova, R. and Jezewski, J. and Leski, J. M.},
-   title = {Detection of Atrial Fibrillation Episodes in Long-Term Heart Rhythm Signals Using a Support Vector Machine},
-   journal = {Sensors (Basel)},
-   volume = {20},
-   number = {3},
-   note = {Czabanski, Robert
-Horoba, Krzysztof
-Wrobel, Janusz
-Matonia, Adam
-Martinek, Radek
-Kupka, Tomasz
-Jezewski, Michal
-Kahankova, Radana
-Jezewski, Janusz
-Leski, Jacek M
-eng
-2017/27/B/ST6/01989/Narodowe Centrum Nauki
-STRATEGMED2/269343/18/ NCBR/2016/Narodowe Centrum Badan i Rozwoju
-Switzerland
-2020/02/06
-Sensors (Basel). 2020 Jan 30;20(3). pii: s20030765. doi: 10.3390/s20030765.},
-   abstract = {Atrial fibrillation (AF) is a serious heart arrhythmia leading to a significant increase of the risk for occurrence of ischemic stroke. Clinically, the AF episode is recognized in an electrocardiogram. However, detection of asymptomatic AF, which requires a long-term monitoring, is more efficient when based on irregularity of beat-to-beat intervals estimated by the heart rate (HR) features. Automated classification of heartbeats into AF and non-AF by means of the Lagrangian Support Vector Machine has been proposed. The classifier input vector consisted of sixteen features, including four coefficients very sensitive to beat-to-beat heart changes, taken from the fetal heart rate analysis in perinatal medicine. Effectiveness of the proposed classifier has been verified on the MIT-BIH Atrial Fibrillation Database. Designing of the LSVM classifier using very large number of feature vectors requires extreme computational efforts. Therefore, an original approach has been proposed to determine a training set of the smallest possible size that still would guarantee a high quality of AF detection. It enables to obtain satisfactory results using only 1.39% of all heartbeats as the training data. Post-processing stage based on aggregation of classified heartbeats into AF episodes has been applied to provide more reliable information on patient risk. Results obtained during the testing phase showed the sensitivity of 98.94%, positive predictive value of 98.39%, and classification accuracy of 98.86%.},
-   keywords = {Algorithms
-Atrial Fibrillation/*diagnosis/physiopathology
-Databases, Factual
-Diagnosis, Computer-Assisted
-Electrocardiography/*methods
-Heart Rate/*physiology
-Humans
-Signal Processing, Computer-Assisted
-Support Vector Machine
-AF detection
-HRV features
-atrial fibrillation (AF)
-heart rate variability (HRV)
-support vector machine (SVM)},
-   ISSN = {1424-8220 (Electronic)
-1424-8220 (Linking)},
-   DOI = {10.3390/s20030765},
-   url = {https://www.ncbi.nlm.nih.gov/pubmed/32019220},
-   year = {2020}
-}
-
-@article{RN99,
-   author = {David, Garner and Naiara, Souza and Luiz, Vanderlei},
-   title = {- Heart Rate Variability Analysis: Higuchi and Katz’s Fractal Dimensions in Subjects with Type 1 Diabetes Mellitus},
-   journal = {Romanian Journal of Diabetes Nutrition and Metabolic Diseases},
-   volume = {25},
-   number = {3},
-   pages = {289-295},
[…]
