### 2023-05-30 #5391 @Kevin-Mattheus-Moerman
@bmcfee thanks for editing here. I recommend the new functionality `@editorialbot generate post review checklist`, as you can see it generates a list of final check. In this case the title of the archive doesn't match yet so we'll ask the authors to fix that.
