+++ paper/paper.bib
@@ -26,7 +26,7 @@ Reproducibility of Results},
 @article{RN88,
    author = {Bohanon, F. J. and Mrazek, A. A. and Shabana, M. T. and Mims, S. and Radhakrishnan, G. L. and Kramer, G. C. and Radhakrishnan, R. S.},
    title = {Heart rate variability analysis is more sensitive at identifying neonatal sepsis than conventional vital signs},
-   journal = {AMERICAN JOURNAL OF SURGERY},
+   journal = {American Journal of Surgery},
    volume = {210},
    number = {4},
    pages = {661-667},
@@ -142,7 +142,7 @@ Reproducibility of Results},
 @article{RN90,
    author = {De Couck, M. and Gidron, Y.},
    title = {Norms of vagal nerve activity, indexed by Heart Rate Variability, in cancer patients},
-   journal = {CANCER EPIDEMIOLOGY},
+   journal = {Cancer Epidemiology},
    volume = {37},
    number = {5},
    pages = {737-741},
@@ -187,7 +187,7 @@ sport},
 @article{RN97,
    author = {Giannakakis, G. and Tsiknakis, M. and Vorgia, P.},
    title = {Focal epileptic seizures anticipation based on patterns of heart rate variability parameters},
-   journal = {COMPUTER METHODS AND PROGRAMS IN BIOMEDICINE},
+   journal = {Computer Methods and Programs in Biomedicine},
    volume = {178},
    pages = {123-132},
    abstract = {Background and Objective: Heart rate variability parameters are studied by the research community as potential valuable indices for seizure detection and anticipation. This paper investigates heart activity abnormalities during focal epileptic seizures in childhood. Methods: Seizures affect both the sympathetic and parasympathetic system which is expressed as abnormal patterns of heart rate variability (HRV) parameters. In the present study, a clinical dataset containing 42 focal seizures in long-term electrocardiographic (ECG) recordings from drug-resistant pediatric epileptic patients (with age 8.2 +/- 4.3 years) was analyzed. Results: Results indicate that the time domain HRV parameters (heart rate, SDNN, standard deviation of heart rate, upper envelope) and spectral HRV parameters (LF/HF, normalized HF, normalized LF, total power) are significantly affected during ictal periods. The HRV features were ranked in terms of their relevance and efficacy to discriminate non-ictal/ictal periods and the top-ranked features were selected using the minimum Redundancy Maximum Relevance algorithm for further analysis. Then, a personalized anticipation algorithm based on multiple regression was introduced providing an "epileptic index" of imminent seizures. The performance of the system resulted in anticipation accuracy of 77.1% and an anticipation time of 21.8 s. Conclusions: The results of this analysis could permit the anticipation of focal seizures only using electrocardiographic signals and the implementation of seizure anticipation strategies for a range of real-life clinical applications. (C) 2019 Elsevier B.V. All rights reserved.},
@@ -200,7 +200,7 @@ sport},
 @article{RN95,
    author = {Kamal, S. M. and Babini, M. H. and Krejcar, O. and Namazi, H.},
    title = {Complexity-Based Decoding of the Coupling Among Heart Rate Variability (HRV) and Walking Path},
-   journal = {FRONTIERS IN PHYSIOLOGY},
+   journal = {Frontiers in Physiology},
    volume = {11},
[…]
