### 2023-05-31 486dacb
Updated License to GNU AGPL V3
