### 2023-05-31 #5391 @charlienew
Hi @Kevin-Mattheus-Moerman, thanks for your feedback. I have updated the bib file accordingly.

The release we created was V1.0.0 on our GitHub repo, and the DOI for our Zenodo archive is [10.5281/zenodo.7963494](https://zenodo.org/record/7963494). The title and the license for our archive have also been updated. Thanks.

### Additional Author Tasks After Review is Complete

- [x] Double-check authors and affiliations (including ORCIDs)
